Gate runtime critical flow explicitly in CI
The critical runtime inference harness now has a first-class task and a dedicated CI matrix so Deno, Node, and Bun failures produce stable required-check names instead of hiding under broader runtime suites. The harness also imports a narrow runtime helper seam so affected-file checking no longer traverses the template E2E catalog graph.

Constraint: CI must expose tests (runtime critical flow: deno|node|bun) and run one TypeScript task with one runtime argument per lane
Constraint: Preserve per-lane build isolation until wall-clock data justifies artifact choreography
Rejected: Add test-config aliases for template-runtime-e2e coupling | would hide the broad helper dependency instead of repairing the seam
Confidence: high
Scope-risk: moderate
Directive: Keep this gate dedicated; do not fold it back into broad runtime suites unless required checks are preserved
Tested: deno fmt/check on changed files; deno lint on changed script files; deno check --config=scripts/test.deno.json scripts/test/runtime-inference-critical-flow.ts scripts/test/runtime-inference-critical-flow.test.ts; focused runtime inference contract test; workflow contract bundle; new task smoke invocation
Not-tested: Full three-lane runtime journey rerun; Task 3 already covered the runtime behavior. deno task test:scripts still fails on unrelated scripts/docs/generate-api-reference.test.ts RuntimeMetadata source-anchor assertion
