+++ .github/workflows/cicd.yml
@@ -135,6 +135,34 @@ jobs:
       - name: Run Bun runtime suite
         run: deno task test:bun
 
+  tests-runtime-critical-flow:
+    if: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == github.repository }}
+    runs-on: ubuntu-latest
+    timeout-minutes: 20
+    name: "tests (runtime critical flow: ${{ matrix.runtime }})"
+    strategy:
+      fail-fast: false
+      matrix:
+        runtime: [deno, node, bun]
+    steps:
+      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
+        with:
+          persist-credentials: false
+      - uses: ./.github/actions/setup-deno
+        timeout-minutes: 5
+        with:
+          warm-cache: "true"
+      - uses: actions/setup-node@820762786026740c76f36085b0efc47a31fe5020 # v6
+        with:
+          node-version: "24"
+          package-manager-cache: false
+      - uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
+        if: ${{ matrix.runtime == 'bun' }}
+        with:
+          bun-version: "1.3.6"
+      - name: Run runtime critical flow
+        run: deno task test:e2e:runtime-inference-critical-flow --runtime=${{ matrix.runtime }}
+
   coverage-shards:
     if: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == github.repository }}
     runs-on: ubuntu-latest
